Created a simple page to use Jmetal
There is simple page created to use Jmetal.
There is a one funtional page create (ExperiementPage), in this page you are able to define a experiment to run is a see a simple result.
The only experiment that you define now is a experiment using as algorithm  GenerationalGeneticAlgorithm in both form, as problem using OneMax, as crossoverOperator using SinglePointCrossover, as mutationOperator using BitFlipMutation and as SelectionOperator using BinaryTournamentSelection.