Added charts to show results
There is two grafics one is live grafic to show the result while the experiment is running and the other is to show all result of that experiment.
The first one  don't show all results, shows some of first and the last results apart from some intermediate if these have been different from the first.
The second grafic is attached to the button to show all result although they are all the same.

The data is collected from mySql database which is dockerized.